wip(test): healing, friendship, and pp recovery tests
